Break down SignIn component into LogIn and SignUp components
